src/requirements.txt
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
colorama==0.4.6
cryptography==42.0.1
dnspython==2.4.2
exceptiongroup==1.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
kaitaistruct==0.10
multipledispatch==1.0.0
outcome==1.3.0.post0
pyasn1==0.5.1
pycparser==2.21
pymongo==4.6.1
pyOpenSSL==24.0.0
pyparsing==3.1.1
PySocks==1.7.1
PyYAML==6.0.1
selenium==4.17.2
selenium-wire==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
wsproto==1.2.0
zstandard==0.22.0
